Move implementation to separate folder.
