+++ src/impl/nvapi_adapter.cpp
