+++ src/impl/nvapi_adapter_registry.cpp
