+++ src/impl/nvapi_output.cpp
