+++ src/meson.build
@@ -1,8 +1,8 @@
 nvapi_src = files([
   'dxvk/dxvk_interfaces.cpp',
-  'nvapi_output.cpp',
-  'nvapi_adapter.cpp',
-  'nvapi_adapter_registry.cpp',
+  'impl/nvapi_output.cpp',
+  'impl/nvapi_adapter.cpp',
+  'impl/nvapi_adapter_registry.cpp',
   'nvapi_interface.cpp',
 ])
